Found way to send at least 1 million signals

+++ client.c
@@ -1,33 +1,27 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   client.c                                           :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: we <we@student.42.fr>                      +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/03/04 14:49:36 by tjun-yu           #+#    #+#             */
-/*   Updated: 2024/03/26 13:00:09 by we               ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include <unistd.h>
 #include <signal.h>
 #include "libft/Libft.h"
 
-void	receive(int sig, siginfo_t *info, void *ucontext);
+int boolean = 1;
+
+void	receive(int sig, siginfo_t *info, void *ucontext_t);
 
 int	main(int argc, char *argv[])
 {
 	ft_printf("Client PID: %d\n", getpid());
 
-	struct sigaction	action;
-	action.sa_sigaction = receive;
-	action.sa_flags = SA_SIGINFO;
-	sigaction(SIGUSR1, &action, NULL);
+	struct sigaction	action1;
+	action1.sa_sigaction = receive;
+	action1.sa_flags = SA_SIGINFO;
+	sigaction(SIGUSR1, &action1, NULL);
 
+	// Send the first signal
 	kill(ft_atoi(argv[1]), SIGUSR1);
-	while (1)
+
+	// Wait for acknowledgments
+	while (boolean)
+	{
 		pause();
+	}
 
 	ft_printf("Client exit\n");
 
@@ -39,14 +33,14 @@ void	receive(int sig, siginfo_t *info, void *ucontext)
 {
 	static int i;
 
-	ft_printf("Signal[%d] received\n", ++i);
-	ft_printf("Server PID: %d\n", info->si_pid);
-
-	for (int i = 0; i < 10000; i++)
-	{
+	if (++i <= 1000000) {
+		// Send the next signal
 		kill(info->si_pid, SIGUSR1);
+		ft_printf("Signal[%d] received: %d | Server PID: %d\n", i, sig, info->si_pid);
+		return;
 	}
+	// ft_printf("PID: %d\n", getpid());
+	boolean = 0;
 
-	(void)sig;
 	(void)ucontext;
 }
